fix: some attributes are moved to the main process Interaction with a third-party s ystem has been started and fixed

+++ internal/processes/server/db/migrations/000002_create_order_table.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE orders (
    id serial primary key,
    identifier varchar not null unique,
    client integer NOT NULL REFERENCES users(id),
-   accrual integer,
+   accrual REAL,
    points_write_off REAL,
    status varchar not null,
    createddate TIMESTAMP default now()
